Access Control

+++ app/Http/Controllers/PostsController.php
@@ -92,6 +92,11 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        // CHeck for correct user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
@@ -126,6 +131,12 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        
+        //Check correct user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+        
         $post->delete();
         
         return redirect('/posts')->with('success', 'Post Removed');
